feat: introduce shared search models
